Fix DM unlock procedure
